Move build_spa_base_url into utils module ready for non-releases uses
